+++ kythe/java/com/google/devtools/kythe/analyzers/java/JavaEntrySets.java
@@ -34,7 +34,6 @@ import com.sun.tools.javac.code.Symbol.ClassSymbol;
 import com.sun.tools.javac.code.Symbol.MethodSymbol;
 import com.sun.tools.javac.code.Symbol.PackageSymbol;
 import com.sun.tools.javac.tree.JCTree;
-import com.sun.tools.javac.util.Name;
 
 import java.util.HashMap;
 import java.util.HashSet;
@@ -44,6 +43,7 @@ import java.util.Set;
 
 import javax.lang.model.element.ElementKind;
 import javax.lang.model.element.Modifier;
+import javax.lang.model.element.Name;
 import javax.tools.JavaFileObject;
 
 /** Specialization of {@link KytheEntrySets} for Java. */